Service of Login and Site are ready

+++ src/com/bcarsoft/login/service/ServiceLogin.java
@@ -7,6 +7,10 @@ package com.bcarsoft.login.service;
 
 import com.bcarsoft.login.dao.IDAOLogin;
 import com.bcarsoft.login.model.Login;
+import com.bcarsoft.struct.singleton.SingEmailUtil;
+import com.bcarsoft.struct.singleton.SingPassUtil;
+import com.bcarsoft.struct.singleton.SingUserUtil;
+import com.bcarsoft.struct.utilities.StringUtil;
 import java.util.List;
 
 /**
@@ -15,6 +19,7 @@ import java.util.List;
  */
 public class ServiceLogin implements IServiceLogin {
     private final static IDAOLogin DAO;
+    private boolean done;
     
     static {DAO = new com.bcarsoft.login.dao.DAOLogin();}
 
@@ -25,6 +30,16 @@ public class ServiceLogin implements IServiceLogin {
      */
     @Override
     public boolean saveLogin(Login login) {
+        // checks fk
+        this.setDone(login.getFk() > 0);
+        if (!this.isDone()) return this.isDone();
+        // checks size
+        this.setDone(this.stringLenChecks(login));
+        if (!this.isDone()) return this.isDone();
+        // checks enter info
+        this.setDone(this.enterInfoCheck(login));
+        if (!this.isDone()) return this.isDone();
+        // if came here success with checks
         return getDAO().saveLogin(login);
     }
 
@@ -35,6 +50,19 @@ public class ServiceLogin implements IServiceLogin {
      */
     @Override
     public boolean updateLogin(Login login) {
+        // checks id
+        this.setDone(login.getId() > 0);        
+        if (!this.isDone()) return this.isDone();
+        // checks fk
+        this.setDone(login.getFk() > 0);
+        if (!this.isDone()) return this.isDone();
+        // checks size
+        this.setDone(this.stringLenChecks(login));
+        if (!this.isDone()) return this.isDone();
+        // checks enter info
+        this.setDone(this.enterInfoCheck(login));
+        if (!this.isDone()) return this.isDone();
+        // if came here success with checks
         return getDAO().updateLogin(login);
     }
 
@@ -45,7 +73,24 @@ public class ServiceLogin implements IServiceLogin {
      */
     @Override
     public boolean deleteLogin(Login login) {
+        // checks id
+        this.setDone(login.getId() > 0);
+        if (!this.isDone()) return this.isDone();
+        // if came here, success with checks
         return getDAO().deleteLogin(login);
+    }
+    
+    /**
+     * Delete all logins by the foreign key.
+     * @param login Login instance.
+     * @return true if success.
+     */
+    @Override
+    public boolean deleteAllLogin(Login login) {
+        this.setDone(login.getFk() > 0);
+        if (!this.isDone()) return this.isDone();
+        // if came here, success with checks
+        return getDAO().deleteAllLogin(login);
     }
 
     /**
@@ -55,6 +100,9 @@ public class ServiceLogin implements IServiceLogin {
      */
     @Override
     public List<Login> findAllLogin(Login login) {
+        this.setDone(login.getFk() > 0);
+        if (!this.isDone()) return null;
+        // if came here, success with checks
         return getDAO().findAllLogin(login);
     }
 
@@ -66,13 +114,64 @@ public class ServiceLogin implements IServiceLogin {
      */
     @Override
     public List<Login> findSpecificLogin(List data, String sql) {
+        if (data == null || StringUtil.isNullOrEmpty(sql)) return null;
+        // if came here, checks success
         return getDAO().findSpecificLogin(data, sql);
+    }
+    
+    /**
+     * This method can verify if a string has a valid size.
+     * @param login Login instance.
+     * @return true if success.
+     */
+    private boolean stringLenChecks(Login login) {
+        this.setDone(StringUtil.isNullOrEmpty(login.getSiteLogin()));
+        if (this.isDone()) return !this.isDone();
+        this.setDone(StringUtil.isNullOrEmpty(login.getUrlSiteLogin()));
+        if (this.isDone()) return !this.isDone();
+        this.setDone(StringUtil.isNullOrEmpty(login.getUserLogin()));
+        if (this.isDone()) return !this.isDone();
+        this.setDone(StringUtil.isNullOrEmpty(login.getEmailLogin()));
+        if (this.isDone()) return !this.isDone();
+        this.setDone(StringUtil.isNullOrEmpty(login.getPassLogin()));
+        if (this.isDone()) return !this.isDone();
+        // if came here, success with checks
+        this.setDone(true);
+        return this.isDone();
+    }
+    
+    /**
+     * This method can checker information like user, pass, email.
+     * @param login Login instance.
+     * @return true if success.
+     */
+    private boolean enterInfoCheck(Login login) {
+        // check user name
+        this.setDone(SingUserUtil.getInstance().isValidUsername(login.getUserLogin()));
+        if (!this.isDone()) return this.isDone();
+        // check email
+        this.setDone(SingEmailUtil.getInstance().isEmailValid(login.getEmailLogin()));
+        if (!this.isDone()) return this.isDone();
+        // check password
+        this.setDone(SingPassUtil.getInstance().isValidPassword(login.getPassLogin()));
+        if (!this.isDone()) return this.isDone();
+        // if came here, checks success
+        this.setDone(true);
+        return this.isDone();
     }
 
     // getters
     
-    public static IDAOLogin getDAO() {
+    private static IDAOLogin getDAO() {
         return DAO;
     }
-    
+
+    protected boolean isDone() {
+        return done;
+    }
+
+    protected void setDone(boolean done) {
+        this.done = done;
+    }
+        
 }
