+++ src/com/bcarsoft/site/service/ServiceSite.java
@@ -7,6 +7,7 @@ package com.bcarsoft.site.service;
 
 import com.bcarsoft.site.dao.IDAOSite;
 import com.bcarsoft.site.model.Site;
+import com.bcarsoft.struct.utilities.StringUtil;
 import java.util.List;
 
 /**
@@ -25,6 +26,11 @@ public class ServiceSite implements IServiceSite {
      */
     @Override
     public boolean saveSite(Site site) {
+        // checks fk
+        if (!(site.getFk() > 0)) return false;
+        // checks data
+        if (this.isNullOrEmptyString(site)) return false;
+        // if came here, success
         return getDAO().saveSite(site);
     }
 
@@ -35,6 +41,13 @@ public class ServiceSite implements IServiceSite {
      */
     @Override
     public boolean updateSite(Site site) {
+        // checks id
+        if (!(site.getId() > 0)) return false;
+        // checks fk
+        if (!(site.getFk() > 0)) return false;
+        // checks data
+        if (this.isNullOrEmptyString(site)) return false;
+        // if came here, success
         return getDAO().updateSite(site);
     }
 
@@ -45,7 +58,19 @@ public class ServiceSite implements IServiceSite {
      */
     @Override
     public boolean deleteSite(Site site) {
+        if (!(site.getId() > 0)) return false;
         return getDAO().deleteSite(site);
+    }
+    
+    /**
+     * This method checks data for Delete All.
+     * @param site Site Instance.
+     * @return true if success and deleted.
+     */
+    @Override
+    public boolean deleteAllSite(Site site) {
+        if (!(site.getFk() > 0)) return false;
+        return getDAO().deleteAllSite(site);
     }
 
     /**
@@ -55,6 +80,8 @@ public class ServiceSite implements IServiceSite {
      */
     @Override
     public List<Site> findAllSite(Site site) {
+        if (site.getFk() < 1) return null;
+        // if came here, checks success
         return getDAO().findAllSite(site);
     }
 
@@ -66,7 +93,20 @@ public class ServiceSite implements IServiceSite {
      */
     @Override
     public List<Site> findSpecificSite(List data, String sql) {
+        // checks the data list and sql string command
+        if (data == null || StringUtil.isNullOrEmpty(sql)) return null;
+        // if came here, success with checks
         return getDAO().findSpecificSite(data, sql);
+    }
+    
+    /**
+     * This method checks if there's empty or null string.
+     * @param site Site Instance.
+     * @return true if success.
+     */
+    private boolean isNullOrEmptyString(Site site) {
+        if (StringUtil.isNullOrEmpty(site.getNameSite())) return true;
+        return StringUtil.isNullOrEmpty(site.getUrlSite());
     }
 
     // getters
